主人公左右移動実装

+++ LampGame/LampGame/Obj_Hero.h
@@ -39,7 +39,7 @@ class CObjHero : public CObj
 		float m_vy;
 		float m_posture;	//姿勢
 
-		int m_ani_time;		//アニメーションフレーム動作間隔
+		float m_ani_time;		//アニメーションフレーム動作間隔
 		int m_ani_frame;	//描画フレーム
 
 		float m_speed_power;	//スピードパワー
